convert to lib

+++ src/api/models.rs
@@ -1,4 +1,4 @@
-use serde::{Deserialize};
+use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize)]
 pub struct LoginResponse {
@@ -10,7 +10,7 @@ pub struct Life360API {
     pub access_token: String,
 }
 
-#[derive(Debug, Deserialize, Clone)]
+#[derive(Debug, Deserialize, Clone, Serialize)]
 pub struct Circle {
     pub id: String,
     pub name: String,
@@ -18,7 +18,7 @@ pub struct Circle {
     pub members: Option<Vec<Member>>,
 }
 
-#[derive(Debug, Deserialize, Clone)]
+#[derive(Debug, Deserialize, Clone, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Member {
     pub id: String,
@@ -27,16 +27,16 @@ pub struct Member {
     pub location: Option<Location>,
 }
 
-#[derive(Debug, Deserialize, Clone)]
+#[derive(Debug, Deserialize, Clone, Serialize)]
 pub struct Location {
     pub latitude: String,
     pub longitude: String,
     pub accuracy: String,
     pub battery: String,
-    pub speed: i32,
+    pub speed: f64,
 }
 
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct CircleResponse {
     pub circles: Vec<Circle>,
 }
